+++ components/composer/src/Composer.svelte
@@ -104,6 +104,8 @@
   export let show_cc_button: Attribute;
   export let show_bcc_button: Attribute;
   export let show_attachment_button: Attribute;
+  export let max_file_size: number;
+  export let show_max_file_size: Attribute;
   export let show_save_as_draft: Attribute;
   export let show_editor_toolbar: Attribute;
   export let theme: string | void;
@@ -130,6 +132,8 @@
     show_cc_button: true,
     show_bcc_button: true,
     show_attachment_button: true,
+    max_file_size: 0,
+    show_max_file_size: true,
     show_save_as_draft: true,
     show_editor_toolbar: true,
     theme: "auto",
@@ -218,6 +222,21 @@
   $: if (value) {
     mergeMessage(value);
   }
+  let maxFileSize: number;
+  $: {
+    if (!uploadFile) {
+      //Using Nylas file upload api, default 4MB max size.
+      maxFileSize =
+        _this.max_file_size &&
+        _this.max_file_size < 4 &&
+        _this.max_file_size > 0
+          ? _this.max_file_size
+          : 4;
+    } else {
+      //Using custom uploadFile function
+      maxFileSize = _this.max_file_size;
+    }
+  }
 
   let isAttachmentLoaded = false;
   $: if (value?.files?.length > 0 && !isAttachmentLoaded) {
@@ -273,8 +292,8 @@
       fileSelector.value = "";
       if (beforeFileUpload) beforeFileUpload(file);
 
-      if (file.size >= 4000000) {
-        throw `Maximum file size is 4MB. Please upload a different file.`;
+      if (maxFileSize > 0 && file.size >= maxFileSize * 1000000) {
+        throw `Maximum file size is ${maxFileSize}MB. Please upload a different file.`;
       }
 
       const result = uploadFile
@@ -661,9 +680,18 @@
   }
   .composer-btn.file-upload,
   .composer-btn.save-draft {
-    margin-left: 10px;
+    margin-left: 6px;
     width: 32px;
     height: 32px;
+    display: flex;
+    justify-content: center;
+    align-items: center;
+  }
+  .file-size {
+    margin-left: 6px;
+    color: var(--composer-icons-color, #666774);
+    height: 32px;
+    width: auto;
     display: flex;
     justify-content: center;
     align-items: center;
@@ -952,17 +980,6 @@
             {#if isPending}Sending{:else}Send{/if}
           </button>
         </div>
-        {#if _this.show_attachment_button && (id || uploadFile)}
-          <label
-            for="file-upload"
-            class="composer-btn file-upload"
-            title="Attach Files (up to 4MB)"
-            tabindex="0"
-          >
-            <AttachmentIcon class="FooterIcon" />
-            <span class="sr-only">Attach Files</span>
-          </label>
-        {/if}
         {#if _this.show_save_as_draft}
           <button
             for="save-draft"
@@ -973,6 +990,22 @@
             <DraftIcon class="FooterIcon" />
             <span class="sr-only">Save Draft</span>
           </button>
+        {/if}
+        {#if _this.show_attachment_button && (id || uploadFile)}
+          <label
+            for="file-upload"
+            class="composer-btn file-upload"
+            title="Attach Files (up to {maxFileSize}MB)"
+            tabindex="0"
+          >
+            <AttachmentIcon class="FooterIcon" />
+            <span class="sr-only">Attach Files</span>
+          </label>
+          {#if _this.show_max_file_size}
+            <div class="file-size">
+              <span>Max file size: {maxFileSize}MB</span>
+            </div>
+          {/if}
         {/if}
 
         <form action="" enctype="multipart/form-data">