+++ components/composer/CHANGELOG.md
@@ -20,6 +20,7 @@
 - Added a prop `focus_body_onload` which allows to set focus on the 'body' field on load [#379](https://github.com/nylas/components/pull/379)
 - Added border-left css to visually differentiate the current message from the previous messages [#379](https://github.com/nylas/components/pull/379)
 - Added button for user to save email message as draft [#383](https://github.com/nylas/components/pull/383)
+- Added a prop `show_max_file_size` to toggle showing the max file size of attachment. [#386](https://github.com/nylas/components/pull/386)
 
 ## Bug Fixes